Gut the RCHTTP class, remove old UUID fetcher, move stuff around.
